feat : create detail screen ui design and searchbar debounce feature.

+++ core/data/src/main/java/com/example/data/local/entity/SatelliteDetailEntity.kt
@@ -9,7 +9,7 @@ import java.io.Serializable
 data class SatelliteDetailEntity(
     @PrimaryKey
     val id: Int,
-    val costPerLaunch: Int,
+    val costPerLaunch: Long,
     val firstFlight: String,
     val height: Int,
     val mass: Int
